# Performance Improvement for `SourceContainsDriveLetter` Method