--- a/TruePath.Benchmarks/DriveLetterBenchmark.cs
+++ b/TruePath.Benchmarks/DriveLetterBenchmark.cs
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2024 TruePath contributors <https://github.com/ForNeVeR/TruePath>
+//
+// SPDX-License-Identifier: MIT
+
 using BenchmarkDotNet.Attributes;
 
 namespace TruePath.Benchmarks;