#### Description

This pull request improves the performance of the `SourceContainsDriveLetter` method by optimizing the drive letter check. The new implementation leverages a more efficient method to verify if the source contains a drive letter, resulting in better performance as indicated by the benchmark results.

#### Benchmark Results

The benchmark results show a significant improvement in performance:

| Method              | Input  | Mean     | Error     | StdDev    |
|-------------------- |------- |---------:|----------:|----------:|
| UseLatinLetterRange | //     | 1.276 ns | 0.0195 ns | 0.0182 ns |
| UseSearchValues     | //     | 1.538 ns | 0.0154 ns | 0.0144 ns |
| UseLatinLetterRange | a://   | 1.524 ns | 0.0157 ns | 0.0147 ns |
| UseSearchValues     | a://   | 1.760 ns | 0.0309 ns | 0.0289 ns |
| UseLatinLetterRange | A://   | 1.539 ns | 0.0237 ns | 0.0222 ns |
| UseSearchValues     | A://   | 1.769 ns | 0.0123 ns | 0.0103 ns |
| UseLatinLetterRange | foobar | 1.293 ns | 0.0113 ns | 0.0094 ns |
| UseSearchValues     | foobar | 1.766 ns | 0.0175 ns | 0.0164 ns |
| UseLatinLetterRange | k://   | 1.535 ns | 0.0137 ns | 0.0122 ns |
| UseSearchValues     | k://   | 1.766 ns | 0.0181 ns | 0.0160 ns |
| UseLatinLetterRange | K://   | 1.540 ns | 0.0196 ns | 0.0183 ns |
| UseSearchValues     | K://   | 1.772 ns | 0.0212 ns | 0.0198 ns |
| UseLatinLetterRange | r://   | 1.537 ns | 0.0196 ns | 0.0183 ns |
| UseSearchValues     | r://   | 1.768 ns | 0.0161 ns | 0.0151 ns |
| UseLatinLetterRange | R://   | 1.544 ns | 0.0201 ns | 0.0178 ns |
| UseSearchValues     | R://   | 1.761 ns | 0.0133 ns | 0.0124 ns |
| UseLatinLetterRange | z://   | 1.541 ns | 0.0147 ns | 0.0137 ns |
| UseSearchValues     | z://   | 1.768 ns | 0.0148 ns | 0.0139 ns |
| UseLatinLetterRange | Z://   | 1.539 ns | 0.0228 ns | 0.0202 ns |
| UseSearchValues     | Z://   | 1.766 ns | 0.0215 ns | 0.0201 ns |

On average, the UseLatinLetterRange method is approximately 13-24% faster than the UseSearchValues method across various input scenarios.